eval/requirements.txt
ragas>=0.1.0
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
pandas>=2.0.0
matplotlib>=3.7.0
datasets>=2.14.0
PyYAML>=6.0
tqdm>=4.65.0
